+++ lib/domain/repositories/live_operation_repository.dart
@@ -1,0 +1,5 @@
+import '../entities/live_operation_snapshot.dart';
+
+abstract interface class LiveOperationRepository {
+  Future<LiveOperationSnapshot> fetchSnapshot({required String parkId});
+}